GC should not collect true and false
usually, nil true and false are safe because they starts at the begin of the oldspace and as a hack, the GC mostly knows to skip them during the collection.

Howver relying on this hack is fragile and may be the root (pun) of subtil bugs.

Morever, because nil, true and false are used a lot, they will be usually marked anyway making such bugs harder to track.

So it is better to just mark them.