Become should not forget to remember
During the prevalidation of a Become operation (`SpurMemoryManager>>#containsOnlyValidBecomeObjects:and:twoWay:copyHash:`) the arrays are traversed to check that objects are legal to become and to ensure there is available memory.
To do so, the elements of the arrays that are forwarders need to be followed. As an optimization, the array are also updated to contains the forward resolution.

Unfortunately, the traversal and the update use low level pointer access instead of high level slot access that know how to handle old->new reference and update the rememberSet.
Therefore, after such low-level updates, a oldspace array can contains pointer to the newspace witout being remembered.

To solve this, I just propose remove the low-lovel broken update of pointers.
It is not an issue because the real Become operation will correctly follow and correctly update pointers.
The only drawback is that any possible following will be done twice.