Add script `scripts/vmmaker` to build VMMaker image
This commit adds a new "script" to build VMMaker image. This script is
using "smalltalk makefiles" [1] mainly to fit into my (Jan's) workflow.

I'm open to rewrite the script in the same way as other scripts in
`scripts` directory and then having the makefile somewhere else if
anyone objects. However for now this will have to do.